Expose OPA log level as OPAL_INLINE_OPA_LOG_LEVEL
