+++ packages/opal-client/opal_client/opa/options.py
@@ -4,7 +4,14 @@ from typing import List, Optional
 from pydantic import BaseModel, Field
 
 
-class LogLevel(str, Enum):
+class OpaLogFormat(str, Enum):
+    NONE = "none"  # no opa logs are piped
+    MINIMAL = "minimal"  # only the event name is logged
+    HTTP = "http"  # tries to extract http method, path and response status code
+    FULL = "full"  # logs the entire data dict returned
+
+
+class OpaLogLevel(str, Enum):
     info = "info"
     debug = "debug"
     error = "error"
@@ -57,7 +64,9 @@ class OpaServerOptions(BaseModel):
     tls_private_key_file: Optional[str] = Field(
         None, description="path of TLS private key file"
     )
-    log_level: LogLevel = Field(LogLevel.info, description="log level for opa logs")
+    log_level: OpaLogLevel = Field(
+        OpaLogLevel.info, description="log level for opa logs"
+    )
     files: Optional[List[str]] = Field(
         None,
         description="list of built-in rego policies and data.json files that must be loaded into OPA on startup. e.g: system.authz policy when using --authorization=basic, see: https://www.openpolicyagent.org/docs/latest/security/#authentication-and-authorization",
